Add more tests for wildcards
Related to #1794.
