+++ framework/src/test/resources/log4j2.xml
@@ -9,7 +9,7 @@
       </Console>
    </Appenders>
    <Loggers>
-      <Logger name="fr.pilato.elasticsearch.crawler.fs.framework" level="info" additivity="false">
+      <Logger name="fr.pilato.elasticsearch.crawler.fs.framework" level="trace" additivity="false">
          <AppenderRef ref="CONSOLE"/>
       </Logger>
       <Logger name="fr.pilato.elasticsearch.crawler.fs.test" level="info" additivity="false">
